Separate Entry from (Entry) Draft
In the initial SAO API design we wanted to merge what is
conceptually defined as an Entry ("submittion" in CMS) and a
Draft ("user-test" in CMS), since both share an almost identical
entity definition into the database, with the only exception that
the latter one is not being referenced by a Result tuple.

Unfortunately, merging Entry and Draft together carried several
compromises in the API implementation. So, I'm basically stepping
back the initial intention.

+++ model/entry.go
@@ -6,18 +6,17 @@ const (
 )
 
 type Entry struct {
-	ID          int64
+	ID          int
 	Kind        int
-	ContestID   int64  `db:"contest_id"`
+	ContestID   int    `db:"contest_id"`
 	ContestSlug string `db:"contest_slug"`
-	TaskID      int64  `db:"task_id"`
+	TaskID      int    `db:"task_id"`
 	TaskSlug    string `db:"task_slug"`
-	ResultID    int64  `db:"result_prtl_id"`
-	ScoreID     int64
+	ResultID    int    `db:"result_prtl_id"`
+	ScoreID     int
 }
 
-
-func (e *Entry) isDryRun () bool {
+func (e *Entry) isDryRun() bool {
 	if e.Kind == KindUserTest {
 		return true
 	}
